# openapi.yaml
openapi: 3.0.3
info:
  title: Cloud-Based Image Upload & Metadata Service
  description: API for managing image uploads to S3 and metadata in DynamoDB.
  version: 1.0.0
servers:
  - url: http://localhost:4566
    description: LocalStack environment
  - url: https://api.example.com/v1
    description: Production environment

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Internal Server Error
        details:
          type: string
          nullable: true

    GenerateUploadUrlRequest:
      type: object
      required:
        - filename
      properties:
        filename:
          type: string
          example: my-vacation-photo.jpg
        filetype:
          type: string
          example: image/jpeg

    GenerateUploadUrlResponse:
      type: object
      properties:
        upload_url:
          type: string
          description: Presigned S3 upload URL
          example: https://s3.amazonaws.com/bucket/key?...
        object_name:
          type: string
          description: Unique identifier/key for the object in S3
          example: 550e8400-e29b-41d4-a716-446655440000-my-vacation-photo.jpg

    MetadataItem:
      type: object
      required:
        - user_id
        - image_id
        - tag
      properties:
        user_id:
          type: string
          example: user_12345
        image_id:
          type: string
          description: ISO timestamp based unique ID matching S3 object key
          example: 2023-10-27T10:00:00Z-photo.jpg
        tag:
          type: string
          example: vacation
        created_at:
          type: string
          format: date-time
        filename:
          type: string
      additionalProperties: true

    SaveMetadataRequest:
      type: object
      required:
        - user_id
        - image_id
        - tag
      properties:
        user_id:
          type: string
          example: user_12345
        image_id:
          type: string
          example: 2023-10-27T10:00:00Z-photo.jpg
        tag:
          type: string
          example: vacation
        filename:
          type: string
          example: photo.jpg

    SaveMetadataResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/MetadataItem'

    ImageListResponse:
      type: object
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/MetadataItem'

    DownloadUrlResponse:
      type: object
      properties:
        download_url:
          type: string
          description: Presigned S3 download URL
          example: https://s3.amazonaws.com/bucket/key?...

    DeleteResponse:
      type: object
      properties:
        status:
          type: string
          example: deleted
        id:
          type: string
          example: 2023-10-27T10:00:00Z-photo.jpg
        errors:
          type: array
          items:
            type: string
          example: ["S3 delete failed"]
        debug:
          type: string

security:
  - ApiKeyAuth: []

tags:
  - name: Uploads
    description: Operations related to file uploads
  - name: Metadata
    description: Operations related to image metadata
  - name: Downloads
    description: Operations related to file downloads

paths:
  /generate-upload-url:
    post:
      tags:
        - Uploads
      summary: Get presigned upload URL
      description: Generates a presigned S3 URL for uploading an image.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateUploadUrlRequest'
      responses:
        '200':
          description: Successfully generated URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateUploadUrlResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /save-metadata:
    post:
      tags:
        - Metadata
      summary: Save image metadata
      description: Stores metadata for an uploaded image in DynamoDB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveMetadataRequest'
      responses:
        '201':
          description: Metadata saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveMetadataResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /images:
    get:
      tags:
        - Metadata
      summary: List images
      description: Retrieve a list of images filtered by user, tag, or date range.
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
          description: Filter by User ID
        - in: query
          name: tag
          schema:
            type: string
          description: Filter by Tag
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
          description: Start of date range (ISO 8601)
        - in: query
          name: end_date
          schema:
            type: string
            format: date-time
          description: End of date range (ISO 8601)
      responses:
        '200':
          description: List of images matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageListResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /generate-download-url:
    get:
      tags:
        - Downloads
      summary: Get presigned download URL
      description: Generates a presigned S3 URL for downloading an image.
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The Image ID (object key) to download
      responses:
        '200':
          description: Successfully generated URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'
        '400':
          description: Missing Image ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /delete:
    delete:
      tags:
        - Uploads
        - Metadata
      summary: Delete image and metadata
      description: Deletes an image from S3 and its metadata from DynamoDB. Handles partial failures idempotently.
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The Image ID to delete
        - in: query
          name: user_id
          schema:
            type: string
          required: true
          description: The User ID who owns the image (for DynamoDB Key)
      responses:
        '200':
          description: Successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '207':
          description: Partial success (e.g., S3 deleted but DB failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'